# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mariela Martínez Inmobiliaria API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PropertyType = typeof PropertyType[keyof typeof PropertyType];


export const PropertyType = {
  casa: 'casa',
  departamento: 'departamento',
  terreno: 'terreno',
  local: 'local',
  oficina: 'oficina',
  campo: 'campo',
} as const;

export type PropertyOperation = typeof PropertyOperation[keyof typeof PropertyOperation];


export const PropertyOperation = {
  venta: 'venta',
  alquiler: 'alquiler',
} as const;

export type PropertyCurrency = typeof PropertyCurrency[keyof typeof PropertyCurrency];


export const PropertyCurrency = {
  ARS: 'ARS',
  USD: 'USD',
} as const;

export interface Property {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  type: PropertyType;
  operation: PropertyOperation;
  price: number;
  currency: PropertyCurrency;
  neighborhood: string;
  address: string;
  city: string;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  coveredArea?: number | null;
  /** @nullable */
  totalArea?: number | null;
  /** @nullable */
  garage?: boolean | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  additionalFeatures?: string | null;
  photos: string[];
  /** @nullable */
  coverPhoto?: string | null;
  featured: boolean;
  active: boolean;
  createdAt: string;
}

export type PropertyInputType = typeof PropertyInputType[keyof typeof PropertyInputType];


export const PropertyInputType = {
  casa: 'casa',
  departamento: 'departamento',
  terreno: 'terreno',
  local: 'local',
  oficina: 'oficina',
  campo: 'campo',
} as const;

export type PropertyInputOperation = typeof PropertyInputOperation[keyof typeof PropertyInputOperation];


export const PropertyInputOperation = {
  venta: 'venta',
  alquiler: 'alquiler',
} as const;

export type PropertyInputCurrency = typeof PropertyInputCurrency[keyof typeof PropertyInputCurrency];


export const PropertyInputCurrency = {
  ARS: 'ARS',
  USD: 'USD',
} as const;

export interface PropertyInput {
  title: string;
  description?: string;
  type: PropertyInputType;
  operation: PropertyInputOperation;
  price: number;
  currency: PropertyInputCurrency;
  neighborhood: string;
  address: string;
  city: string;
  bedrooms?: number;
  bathrooms?: number;
  coveredArea?: number;
  totalArea?: number;
  garage?: boolean;
  age?: number;
  additionalFeatures?: string;
  photos?: string[];
  coverPhoto?: string;
  featured?: boolean;
  active?: boolean;
}

export type PropertyUpdateType = typeof PropertyUpdateType[keyof typeof PropertyUpdateType];


export const PropertyUpdateType = {
  casa: 'casa',
  departamento: 'departamento',
  terreno: 'terreno',
  local: 'local',
  oficina: 'oficina',
  campo: 'campo',
} as const;

export type PropertyUpdateOperation = typeof PropertyUpdateOperation[keyof typeof PropertyUpdateOperation];


export const PropertyUpdateOperation = {
  venta: 'venta',
  alquiler: 'alquiler',
} as const;

export type PropertyUpdateCurrency = typeof PropertyUpdateCurrency[keyof typeof PropertyUpdateCurrency];


export const PropertyUpdateCurrency = {
  ARS: 'ARS',
  USD: 'USD',
} as const;

export interface PropertyUpdate {
  title?: string;
  description?: string;
  type?: PropertyUpdateType;
  operation?: PropertyUpdateOperation;
  price?: number;
  currency?: PropertyUpdateCurrency;
  neighborhood?: string;
  address?: string;
  city?: string;
  bedrooms?: number;
  bathrooms?: number;
  coveredArea?: number;
  totalArea?: number;
  garage?: boolean;
  age?: number;
  additionalFeatures?: string;
  photos?: string[];
  coverPhoto?: string;
  featured?: boolean;
  active?: boolean;
}

export interface AgencyStats {
  totalProperties: number;
  activeProperties: number;
  forSale: number;
  forRent: number;
  yearsExperience: number;
  propertiesSold: number;
  happyClients: number;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface LoginResult {
  token: string;
}

export interface UploadResult {
  url: string;
}

export interface DeleteResult {
  success: boolean;
}

export type ListPropertiesParams = {
operation?: ListPropertiesOperation;
type?: ListPropertiesType;
neighborhood?: string;
bedrooms?: number;
minPrice?: number;
maxPrice?: number;
featured?: boolean;
active?: boolean;
limit?: number;
offset?: number;
};

export type ListPropertiesOperation = typeof ListPropertiesOperation[keyof typeof ListPropertiesOperation];


export const ListPropertiesOperation = {
  venta: 'venta',
  alquiler: 'alquiler',
} as const;

export type ListPropertiesType = typeof ListPropertiesType[keyof typeof ListPropertiesType];


export const ListPropertiesType = {
  casa: 'casa',
  departamento: 'departamento',
  terreno: 'terreno',
  local: 'local',
  oficina: 'oficina',
  campo: 'campo',
} as const;