# lib/api-zod/src/generated/types/agencyStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mariela Martínez Inmobiliaria API
 * OpenAPI spec version: 0.1.0
 */

export interface AgencyStats {
  totalProperties: number;
  activeProperties: number;
  forSale: number;
  forRent: number;
  yearsExperience: number;
  propertiesSold: number;
  happyClients: number;
}