# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mariela Martínez Inmobiliaria API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyCurrency } from './propertyCurrency';
import type { PropertyOperation } from './propertyOperation';
import type { PropertyType } from './propertyType';

export interface Property {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  type: PropertyType;
  operation: PropertyOperation;
  price: number;
  currency: PropertyCurrency;
  neighborhood: string;
  address: string;
  city: string;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  coveredArea?: number | null;
  /** @nullable */
  totalArea?: number | null;
  /** @nullable */
  garage?: boolean | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  additionalFeatures?: string | null;
  photos: string[];
  /** @nullable */
  coverPhoto?: string | null;
  featured: boolean;
  active: boolean;
  createdAt: string;
}