# lib/api-zod/src/generated/types/propertyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mariela Martínez Inmobiliaria API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyInputCurrency } from './propertyInputCurrency';
import type { PropertyInputOperation } from './propertyInputOperation';
import type { PropertyInputType } from './propertyInputType';

export interface PropertyInput {
  title: string;
  description?: string;
  type: PropertyInputType;
  operation: PropertyInputOperation;
  price: number;
  currency: PropertyInputCurrency;
  neighborhood: string;
  address: string;
  city: string;
  bedrooms?: number;
  bathrooms?: number;
  coveredArea?: number;
  totalArea?: number;
  garage?: boolean;
  age?: number;
  additionalFeatures?: string;
  photos?: string[];
  coverPhoto?: string;
  featured?: boolean;
  active?: boolean;
}