# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mariela Martínez Inmobiliaria API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyUpdateCurrency } from './propertyUpdateCurrency';
import type { PropertyUpdateOperation } from './propertyUpdateOperation';
import type { PropertyUpdateType } from './propertyUpdateType';

export interface PropertyUpdate {
  title?: string;
  description?: string;
  type?: PropertyUpdateType;
  operation?: PropertyUpdateOperation;
  price?: number;
  currency?: PropertyUpdateCurrency;
  neighborhood?: string;
  address?: string;
  city?: string;
  bedrooms?: number;
  bathrooms?: number;
  coveredArea?: number;
  totalArea?: number;
  garage?: boolean;
  age?: number;
  additionalFeatures?: string;
  photos?: string[];
  coverPhoto?: string;
  featured?: boolean;
  active?: boolean;
}